Added the actual tic tac toe board plus lobby improvements
The menu to select a new room now vanishes upon joining one room. The user either has to leave the room or disconnect (refresh the page) to see it again and join a new one.
Once there are 2 players in a room, it will switch to the tic tac toe board so the actual game can happen. The rooms list is now populated locally on the new client that joined.